+++ DarrenChan_KevinDicoy_LenicZhang/.idea/modules.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/DarrenChan_KevinDicoy_LenicZhang.iml" filepath="$PROJECT_DIR$/DarrenChan_KevinDicoy_LenicZhang.iml" />
+      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+    </modules>
+  </component>
+</project>
