+++ DarrenChan_KevinDicoy_LenicZhang/app/proguard-rules.pro
