+++ DarrenChan_KevinDicoy_LenicZhang/app/src/main/java/ca/bcit/ass1/darrenchan_kevindicoy_leniczhang/MainActivity.java
@@ -1,0 +1,13 @@
+package ca.bcit.ass1.darrenchan_kevindicoy_leniczhang;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+    }
+}
