+++ DarrenChan_KevinDicoy_LenicZhang/app/src/main/res/values/colors.xml
