+++ DarrenChan_KevinDicoy_LenicZhang/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Fri Sep 22 13:15:10 PDT 2017
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-3.3-all.zip
