--- MyApplication/app/src/main/java/ca/bcit/ass1/darrenchan_leniczhang_kevindicoy/MainActivity.java
@@ -1,13 +1,0 @@
-package ca.bcit.ass1.darrenchan_leniczhang_kevindicoy;
-
-import android.app.Activity;
-import android.os.Bundle;
-
-public class MainActivity extends Activity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-}
